requirements.txt
backpack==0.1
bcrypt==3.1.3
blinker==1.4
cffi==1.10.0
cleo==0.6.0
click==6.7
enum-compat==0.0.2
eventlet==0.21.0
Faker==0.7.17
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
flask-orator==0.2.0
Flask-SocketIO==2.9.0
Flask-WTF==0.14.2
greenlet==0.4.12
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
orator==0.9.7
pastel==0.1.0
pendulum==1.2.4
psutil==5.2.2
pyaml==16.12.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.4.2
pylev==1.3.0
PyMySQL==0.7.11
python-dateutil==2.6.1
python-engineio==1.7.0
python-socketio==1.7.6
pytz==2017.2
pytzdata==2017.2
PyYAML==3.12
simplejson==3.11.1
six==1.10.0
tzlocal==1.4
Werkzeug==0.12.2
wrapt==1.10.10
WTForms==2.1
WTForms-JSON==0.3.3
